add authChecker for patch and delete

+++ src/controllers/movie.js
@@ -1,7 +1,7 @@
 import * as Response from "../helpers/response/response";
 import Errors from "../helpers/constants/constants";
 import validator from "../validator/movie";
-import getToken from "../helpers/auth/auth";
+import { getToken, checkAuth } from "../helpers/auth/auth";
 import Query from "../database/queries/query";
 
 class MovieData {
@@ -43,19 +43,21 @@ class MovieData {
       return Response.responseServerError(res);
     }
   }
-  //check for authorization before deleting and updating can use a middlware to check if logged in
   static async deleteMovie(req, res) {
     const { id } = req.params;
     try {
-      const { error } = validator.validate({ id });
-      if (error) {
-        return Response.responseValidationError(res, Errors.INVALID_ID);
+      const isAuthorized = checkAuth(req);
+      if (isAuthorized) {
+        const { error } = validator.validate({ id });
+        if (error) {
+          return Response.responseValidationError(res, Errors.INVALID_ID);
+        }
+        const movieToDelete = await Query.deleteMovie(id);
+        if (!movieToDelete) {
+          return Response.responseNotFound(res, Errors.INVALID_MOVIE);
+        }
+        return Response.responseOk(res, movieToDelete);
       }
-      const movieToDelete = await Query.deleteMovie(id);
-      if (!movieToDelete) {
-        return Response.responseNotFound(res, Errors.INVALID_MOVIE);
-      }
-      return Response.responseOk(res, movieToDelete);
     } catch (error) {
       return Response.responseServerError(res);
     }
@@ -65,19 +67,22 @@ class MovieData {
     const userId = getToken(req);
     const movieData = { ...req.body, userId };
     try {
-      const { error } = validator.validate(movieData);
-      if (error) {
-        return Response.responseBadRequest(res, Errors.VALIDATION);
+      const isAuthorized = checkAuth(req);
+      if (isAuthorized) {
+        const { error } = validator.validate(movieData);
+        if (error) {
+          return Response.responseBadRequest(res, Errors.VALIDATION);
+        }
+        const { err } = validator.validate({ id });
+        if (err) {
+          return Response.responseValidationError(res, Errors.INVALID_ID);
+        }
+        const movieToUpdate = await Query.updateMovie(id, movieData);
+        if (movieToUpdate.length == 0) {
+          return Response.responseNotFound(res, Errors.INVALID_MOVIE);
+        }
+        return Response.responseOk(res, movieToUpdate);
       }
-      const { err } = validator.validate({ id });
-      if (err) {
-        return Response.responseValidationError(res, Errors.INVALID_ID);
-      }
-      const movieToUpdate = await Query.updateMovie(id, movieData);
-      if (movieToUpdate.length == 0) {
-        return Response.responseNotFound(res, Errors.INVALID_MOVIE);
-      }
-      return Response.responseOk(res, movieToUpdate);
     } catch (error) {
       return Response.responseServerError(res);
     }
